Refactor : Email, Home, Image 리팩토링
