+++ src/main/java/com/mypill/domain/email/emailsender/service/EmailSenderService.java
@@ -1,4 +1,4 @@
-package com.mypill.domain.emailsender.service;
+package com.mypill.domain.email.emailsender.service;
 
 public interface EmailSenderService {
     void send(String to, String from, String title, String body);
