--- src/main/java/com/mypill/domain/image/repository/ImageRepository.java
@@ -1,8 +1,0 @@
-package com.mypill.domain.image.repository;
-
-import com.mypill.domain.image.entity.Image;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ImageRepository extends JpaRepository<Image, Long> {
-
-}
